third_party/WebKit/LayoutTests/platform/mac/fast/events/touch/gesture/touch-gesture-scroll-listbox-expected.txt


This tests that an input text field can be properly scrolled with touch gestures

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS successfullyParsed is true

TEST COMPLETE
===Testing fling gestures===
PASS box.scrollTop is 0
PASS container.scrollTop is 0
Flinging the list box should scroll the list, the scrolls should be locked to item boundaries
FAIL box.scrollTop should be 38.4. Was 38.
PASS container.scrollTop is 0
Flinging the list past the end shouldn't scroll containing div
PASS box.scrollTop is fullyScrolled
PASS container.scrollTop is 0
Flinging fully scrolled list should fling containing div
PASS box.scrollTop is fullyScrolled
PASS container.scrollTop is 60
===Testing gesture scroll===
PASS box.scrollTop is 0
PASS container.scrollTop is 0
Gesture scrolling list should scroll the list, the scrolls should lock to item boundaries
FAIL box.scrollTop should be 48.599999999999994. Was 48.
PASS container.scrollTop is 0
PASS box.scrollTop is 0
PASS container.scrollTop is 0
Gesture scrolling list past the end shouldn't scroll container div
PASS box.scrollTop is fullyScrolled
PASS container.scrollTop is 0
Gesture scrolling list past the end should scroll container div when starting at scroll extent
PASS box.scrollTop is fullyScrolled
PASS container.scrollTop is fullyScrolled + 50
===Testing horizontal scroll===
PASS box.scrollLeft is 0
PASS container.scrollLeft is 0
Horizontal scrolls should not affect listbox
PASS box.scrollLeft is 0
PASS container.scrollLeft is 60
